pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"
python-requires = ">=3.9.13,<3.12"

[tool.black]
line-length = 88
target-version = ['py36']
include = '\\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[project]
name = "autotransformers"
version = "0.0.1"
description = "a Python package for automatic training and benchmarking of Language Models."
readme = "README.md"
authors = [
    { name = "LenguajeNatural.AI", email = "avaca@lenguajenatural.ai" }
]
keywords = ["large-language-models", "language-models", "natural-language-processing", "nlp", "transformers", "hyperparameter-tuning", "automatic-training"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "datasets>=2.17.1",
    "transformers[torch]>=4.38.1",
    "bitsandbytes>=0.42.0",
    "peft>=0.8.2",
    "trl>=0.7.11",
    "optuna>=3.0.2",
    "scikit-learn>=1.0.2",
    "nltk>=3.7",
    "rouge_score>=0.1.2",
    "tensorboard>=2.10.1",
    "tensorboardX>=2.5.1",
    "sentencepiece>=0.1.97",
    "apscheduler>=3.6.3",
    "seaborn>=0.12.0",
    "nlpaug>=1.1.11",
    "simpletransformers>=0.61.10",
    "pandas>=1.3.5",
    "tqdm>=4.64.1",
    "evaluate>=0.2.2",
    "more-itertools>=8.14.0",
    "polyfuzz>=0.4.0",
    "seqeval>=1.2.2",
    "accelerate>=0.20.3",
    "scipy>=1.11.0",
    "einops>=0.6.1",
    "pytest>=7.1.3"
]



[tool.pytest.ini_options]
addopts = "--cov-report html:cover --cov-report term --cov-config=.coveragerc --junitxml=report.xml --cov=src --doctest-modules --durations=20 --ignore=examples/ --ignore=doc/ --ignore=notebooks/ --instafail --pycodestyle --pydocstyle"
filterwarnings = [
    "ignore::PendingDeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::UserWarning"
]

[tool.pydocstyle]
convention = "numpy"
add-ignore = "D104, D100, D103"

[tool.pycodestyle]
statistics = true
max_line_length = 100
ignore = "W503, W605, E501, E203"

[project.optional-dependencies]

docs = [
    "sphinxemoji>=0.2.0",
    "sphinx>=5.3.0",
    "sphinx-rtd-theme>=1.1.1",
    "readthedocs-sphinx-search>=0.1.2",
    "sphinxcontrib-applehelp>=1.0.2",
    "sphinxcontrib-devhelp>=1.0.2",
    "sphinxcontrib-htmlhelp>=2.0.0",
    "sphinxcontrib-jsmath>=1.0.1",
    "sphinxcontrib-qthelp>=1.0.3",
    "sphinxcontrib-serializinghtml>=1.1.5",
    "numpydoc>=1.5.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = { find = { where = ["src"] } }

[project.urls]
Homepage = "https://lenguajenatural.ai"
repository = "https://github.com/lenguajenatural-ai/autotransformers"